[가원] dateID Long으로 수정, 일기 생성/조회 api 수정완료

+++ drawingdiary/backend/src/main/java/com/diary/drawing/diary/repository/DateRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.diary.drawing.diary.domain.Date;
 
 @Repository
-public interface DateRepository extends JpaRepository<Date, Integer>{
-    Date findByDateID(int dateID); // Optional: findBy로 조회했을 때 객체가 존재하지 않는 경우 처리
+public interface DateRepository extends JpaRepository<Date, Long>{
+    Date findByDateID(Long dateID); // Optional: findBy로 조회했을 때 객체가 존재하지 않는 경우 처리
 }
